requirements.txt
llama-index==0.10.38
llama-index-embeddings-huggingface==0.2.0
llama-index-llms-groq==0.1.4
llama-index-vector-stores-chroma==0.1.8
llama-parse==0.4.3
chainlit==1.1.202
python-dotenv==1.0.1